// app/src/app/store/questions/question-actions-types.ts
export enum questionActionsTypes {
  fetchQuestions = '[Questions] FetchQuestions',
  fetchQuestionsSuccess = '[Questions] FetchQuestionsSuccess',
  fetchQuestionsFail = '[Questions] fetchQuestionsFail',
  newQuestion = '[Questions] NewQuestion',
  newQuestionSuccess = '[Questions] NewQuestionSuccess',
  newQuestionFail = '[Questions] NewQuestionFail',

  saveQuestionAnswer = '[Questions] SaveQuestionAnswer',
  saveQuestionAnswerSuccess = '[Questions] SaveQuestionAnswerSuccess',
  saveQuestionAnswerFail = '[Questions] SaveQuestionAnswerFail',
}